+++ tests/testthat/test-session.R
@@ -56,3 +56,51 @@ test_that("pnf error is validated", {
     "as long as `pkg`"
   )
 })
+
+test_that("can handle check-installed", {
+  local_interactive()
+
+  # Override `is_installed()` results
+  override <- NULL
+  is_installed_hook <- function(pkg, ver) {
+    if (is_bool(override)) {
+      rep_along(pkg, override)
+    } else {
+      with_options(
+        "rlang:::is_installed_hook" = NULL,
+        is_installed(pkg, version = ver)
+      )
+    }
+  }
+  local_options("rlang:::is_installed_hook" = is_installed_hook)
+
+  test_env <- current_env()
+  handle <- function(value, frame, expr) {
+    withCallingHandlers(
+      rlib_error_package_not_found = function(cnd) {
+        override <<- value
+        invokeRestart("rlib_restart_package_not_found")
+      },
+      expr
+    )
+  }
+
+  override <- NULL
+  expect_no_error(
+    handle(
+      TRUE,
+      test_env,
+      check_installed(c("foo", "bar"), version = c("1.0", "2.0"))
+    )
+  )
+
+  override <- NULL
+  expect_error(
+    handle(
+      FALSE,
+      test_env,
+      check_installed(c("foo", "bar"), version = c("1.0", "2.0"))
+    ),
+    "are required"
+  )
+})